LevelLoader
Added script for loading levels, as well as progress bar. Text doesn't update at the moment
