+++ Assets/Audio/Mixers/NewAudioMixer.mixer
@@ -67,5 +67,5 @@ AudioMixerSnapshotController:
   m_AudioMixer: {fileID: 24100000}
   m_SnapshotID: fd2e824785993a245835e2848c37c266
   m_FloatValues:
-    3b9f3a3a090227544a98a90b186d34d0: 0
+    3b9f3a3a090227544a98a90b186d34d0: -32.27831
   m_TransitionOverrides: {}
