+++ Assets/Scenes/Workspaces/Adam Week 9 stuff/Loading Screen.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ebd88c4f75c7c0043a1d6681e492fa22
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
