+++ Assets/Sprites/Game Sprites/Untitled_Artwork.psd.meta
@@ -159,7 +159,9 @@ TextureImporter:
     secondaryTextures: []
     spriteCustomMetadata:
       entries: []
-    nameFileIdTable: {}
+    nameFileIdTable:
+      Untitled_Artwork_0: 1302723820812287101
+      Untitled_Artwork_1: 6146782616825927046
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
